IMP: PC sampler, SJD ablation eval, offline eval, env updates
Predictor-corrector sampler with perturb-check gate, CIFAR-10 SJD
ablation eval config (50k samples), offline evaluation CLI, updated
environment deps, and miscellaneous task/test fixes.

+++ src/sticky/data/openwebtext.py
@@ -121,7 +121,7 @@ def make_openwebtext_iterator(
                 if (stop - start) < int(batch_size) and bool(drop_remainder):
                     break
                 batch_indices = indices[start:stop]
-                batch = np.asarray(tokens[batch_indices], dtype=np.int32)
+                batch = np.asarray(tokens[batch_indices], dtype=np.int64)
                 yield {"image": batch}
 
             if not bool(repeat):
